back/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
pillow
requests
python-dotenv
supabase
better-profanity
